+++ src/KsWare.Presentation.ViewModelFramework/ViewModel/(UI-VM)/(WindowVM)/IWindowVM.cs
@@ -10,6 +10,13 @@
 		void Show();
 
 //		void Close();
+
+		/// <summary>
+		/// Gets or sets the <see cref="IWindowVM">window</see> that owns this <see cref="IWindowVM"/>.
+		/// </summary>
+		/// <value>A <see cref="IWindowVM"/> object that represents the owner of this <see cref="IWindowVM"/>.</value>
+		/// <seealso cref="System.Windows.Window.Owner"/>
+		IWindowVM Owner { get; set; }
 	}
 
 }